requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asn1==2.7.0
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
bitarray==2.6.2
boto3==1.26.59
botocore==1.29.59
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
cytoolz==0.12.1
enum-compat==0.0.3
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
eulith-web3==0.0.12
fireblocks-sdk==1.18.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
hexbytes==0.3.0
httplib2==0.21.0
idna==3.4
ipfshttpclient==0.8.0a2
jmespath==1.0.1
jsonschema==4.17.3
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
oauthlib==3.2.2
parsimonious==0.8.1
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
requests==2.28.2
requests-oauthlib==1.3.1
rlp==2.0.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
toolz==0.12.0
uritemplate==4.1.1
urllib3==1.26.14
uuid==1.30
varint==1.0.2
web3==5.31.3
websockets==9.1
yarl==1.8.2
